requirements.txt
asttokens==2.4.1
async-timeout==5.0.1
beautifulsoup4==4.12.3
blinker==1.7.0
bs4==0.0.2
cachelib==0.12.0
ceiling==0.0.1
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.2
debugpy==1.8.1
decorator==5.1.1
Deprecated==1.2.18
exceptiongroup==1.2.0
executing==2.0.1
Flask==3.0.2
Flask-Cors==4.0.0
Flask-Limiter==3.12
Flask-Session==0.7.0
flask_abort==0.0.1
floor==0.0.1
gevent==24.2.1
greenlet==3.0.3
gunicorn==21.2.0
idna==3.7
ipykernel==6.29.3
ipython==8.22.2
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.3
jsonify==0.5
jupyter_client==8.6.1
jupyter_core==5.7.2
limits==4.4.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.6
mdurl==0.1.2
msgspec==0.18.6
nest-asyncio==1.6.0
numpy==1.26.4
ordered-set==4.1.0
packaging==24.0
pandas==2.2.1
parso==0.8.3
platformdirs==4.2.0
prompt-toolkit==3.0.43
psutil==5.9.8
pure-eval==0.2.2
pycparser==2.21
Pygments==2.17.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
pyzmq==25.1.2
redis==5.2.1
requests==2.31.0
rich==13.9.4
six==1.16.0
soupsieve==2.5
stack-data==0.6.3
tornado==6.4
traitlets==5.14.2
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.1
waitress==3.0.0
wcwidth==0.2.13
Werkzeug==3.0.1
wrapt==1.17.2
zope.event==5.0
zope.interface==6.2
